Version_7v3_D3

+++ 3курс2семестр/Form3.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Numerics;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data;
@@ -17,7 +18,7 @@ namespace _3курс2семестр
             InitializeComponent();
         }
 
-        ulong Factorial(ulong n)
+        BigInteger Factorial(int n)
         {
             if (n == 1) return 1;
 
@@ -28,7 +29,7 @@ namespace _3курс2семестр
         {
             try
             {
-                textBox2.Text = Convert.ToString(Factorial(Convert.ToUInt64(textBox1.Text)));
+                textBox2.Text = Convert.ToString(Factorial(Convert.ToInt32(textBox1.Text)));
             }
             catch (Exception ex)
             {
